Move npm registry parsing into 'registry' module
